% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__drop}
\alias{dataframe__drop}
\title{Drop columns of a DataFrame}
\usage{
dataframe__drop(..., strict = TRUE)
}
\arguments{
\item{...}{Characters of column names to drop. Passed to \code{\link[=pl_col]{pl$col()}}.}

\item{strict}{Validate that all column names exist in the schema and throw an
exception if a column name does not exist in the schema.}
}
\value{
A polars \link{DataFrame}
}
\description{
Drop columns of a DataFrame
}
\examples{
pl$DataFrame(mtcars)$drop(c("mpg", "hp"))

# equivalent
pl$DataFrame(mtcars)$drop("mpg", "hp")
}
