% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__equals}
\alias{dataframe__equals}
\title{Compare two DataFrames}
\usage{
dataframe__equals(other, ..., null_equal = TRUE)
}
\arguments{
\item{other}{DataFrame to compare with.}
}
\value{
A logical value
}
\description{
Check if two DataFrames are equal.
}
\examples{
dat1 <- pl$DataFrame(iris)
dat2 <- pl$DataFrame(iris)
dat3 <- pl$DataFrame(mtcars)
dat1$equals(dat2)
dat1$equals(dat3)
}
