% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__get_column}
\alias{dataframe__get_column}
\title{Get a single column by name}
\usage{
dataframe__get_column(name, ..., default = NULL)
}
\arguments{
\item{name}{Name of the column to retrieve.}

\item{...}{These dots are for future extensions and must be empty.}

\item{default}{Value to return if the column does not exist. If not
explicitly set and the column is not present, an error is thrown.}
}
\value{
Either a Series or a specific value if the column doesn't exist and
\code{default} is set.
}
\description{
Get a single column by name
}
\examples{
df <- pl$DataFrame(foo = 1:3, bar = 4:6)
df$get_column("foo")

# Missing column handling; can optionally provide an arbitrary default value
# to the method (otherwise an error is thrown).
df$get_column("baz", default = pl$Series("baz", c("?", "?", "?")))

df$get_column("baz", default = NA)

tryCatch(
  df$get_column("baz"),
  error = function(e) print(e)
)
}
