% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__select}
\alias{dataframe__select}
\title{Select and modify columns of a DataFrame}
\usage{
dataframe__select(...)
}
\arguments{
\item{...}{Columns to keep. Those can be expressions (e.g \code{pl$col("a")}),
column names  (e.g \code{"a"}), or list containing expressions or column names
(e.g \code{list(pl$col("a"))}).}
}
\value{
A polars \link{DataFrame}
}
\description{
Similar to \code{dplyr::mutate()}. However, it discards unmentioned
columns (like \code{.()} in \code{data.table}).
}
\examples{
pl$DataFrame(iris)$select(
  pl$col("Sepal.Length")$abs()$alias("abs_SL"),
  (pl$col("Sepal.Length") + 2)$alias("add_2_SL")
)
}
