% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__with_columns}
\alias{dataframe__with_columns}
\title{Modify/append column(s)}
\usage{
dataframe__with_columns(...)
}
\arguments{
\item{...}{Any expressions or string column name, or same wrapped in a list.
If first and only element is a list, it is unwrapped as a list of args.}
}
\value{
A polars \link{DataFrame}
}
\description{
Add columns or modify existing ones with expressions. This is
the equivalent of \code{dplyr::mutate()} as it keeps unmentioned columns (unlike
\verb{$select()}).
}
\examples{
as_polars_df(iris)$with_columns(
  abs_SL = pl$col("Sepal.Length")$abs()$alias(""),
  add_2_SL = pl$col("Sepal.Length") + 2$alias("")
)

# same query
l_expr <- list(
  pl$col("Sepal.Length")$abs()$alias("abs_SL"),
  (pl$col("Sepal.Length") + 2)$alias("add_2_SL")
)
as_polars_df(iris)$with_columns(l_expr)

as_polars_df(iris)$with_columns(
  SW_add_2 = (pl$col("Sepal.Width") + 2),
  # unnamed expr will keep name "Sepal.Length"
  pl$col("Sepal.Length")$abs()
)
}
