% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__write_parquet}
\alias{dataframe__write_parquet}
\title{Write to parquet file}
\usage{
dataframe__write_parquet(
  file,
  ...,
  compression = c("lz4", "uncompressed", "snappy", "gzip", "lzo", "brotli", "zstd"),
  compression_level = NULL,
  statistics = TRUE,
  row_group_size = NULL,
  data_page_size = NULL,
  partition_by = NULL,
  partition_chunk_size_bytes = 4294967296,
  storage_options = NULL,
  retries = 2
)
}
\arguments{
\item{file}{File path to which the result should be written. This should be
a path to a directory if writing a partitioned dataset.}

\item{partition_by}{A character vector indicating column(s) to partition by.
A partitioned dataset will be written if this is specified.}

\item{partition_chunk_size_bytes}{Approximate size to split DataFrames within
a single partition when writing. Note this is calculated using the size of
the DataFrame in memory (the size of the output file may differ depending
on the file format / compression).}
}
\value{
Doesn't return anything
}
\description{
Write to parquet file
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat = as_polars_df(mtcars)

# write data to a single parquet file
destination = withr::local_tempfile(fileext = ".parquet")
dat$write_parquet(destination)

# write data to folder with a hive-partitioned structure
dest_folder = withr::local_tempdir()
dat$write_parquet(dest_folder, partition_by = c("gear", "cyl"))
list.files(dest_folder, recursive = TRUE)
\dontshow{\}) # examplesIf}
}
