% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__alias}
\alias{expr__alias}
\title{Rename Expr output}
\usage{
expr__alias(name)
}
\arguments{
\item{name}{New name of output}
}
\value{
A polars \link{expression}
}
\description{
Rename the output of an expression.
}
\examples{
pl$col("bob")$alias("alice")
}
