% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_count}
\alias{expr__cum_count}
\title{Cumulative count}
\usage{
expr__cum_count(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, reverse the count.}
}
\value{
A polars \link{expression}
}
\description{
Get an array with the cumulative count (zero-indexed) computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.

\verb{$cum_count()} does not seem to count within lists.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  cum_count = pl$col("a")$cum_count(),
  cum_count_reversed = pl$col("a")$cum_count(reverse = TRUE)
)
}
