% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_min}
\alias{expr__cum_min}
\title{Cumulative minimum}
\usage{
expr__cum_min(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, start from the last value.}
}
\value{
A polars \link{expression}
}
\description{
Get an array with the cumulative min computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = c(1:4, 2L))$with_columns(
  pl$col("a")$cum_min()$alias("cum_min"),
  pl$col("a")$cum_min(reverse = TRUE)$alias("cum_min_reversed")
)
}
