% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_prod}
\alias{expr__cum_prod}
\title{Cumulative product}
\usage{
expr__cum_prod(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, start with the total product of elements and divide
each row one by one.}
}
\value{
A polars \link{expression}
}
\description{
Get an array with the cumulative product computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  cum_prod = pl$col("a")$cum_prod(),
  cum_prod_reversed = pl$col("a")$cum_prod(reverse = TRUE)
)
}
