% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__ewm_var}
\alias{expr__ewm_var}
\title{Compute exponentially-weighted moving variance}
\usage{
expr__ewm_var(
  ...,
  com,
  span,
  half_life,
  alpha,
  adjust = TRUE,
  bias = FALSE,
  min_periods = 1,
  ignore_nulls = FALSE
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{com}{Specify decay in terms of center of mass.}

\item{span}{Specify decay in terms of span.}

\item{half_life}{Specify decay in terms of half-life.}

\item{alpha}{Specify smoothing factor alpha directly.}

\item{adjust}{Divide by decaying adjustment factor in beginning periods to
account for imbalance in relative weightings:
\itemize{
\item when \code{TRUE} (default), the EW function is calculated using weights
\eqn{w_i = (1 - \alpha)^i};
\item when \code{FALSE}, the EW function is calculated recursively by \deqn{y_0 = x_0}
\deqn{y_t = (1 - \alpha)y_{t - 1} + \alpha x_t}
}}

\item{bias}{If \code{FALSE} (default), apply a correction to make the estimate
statistically unbiased.}

\item{min_periods}{The number of values in the window that should be
non-null before computing a result. If \code{NULL} (default), it will be set
equal to \code{window_size}.}

\item{ignore_nulls}{Ignore missing values when calculating weights.
\itemize{
\item when \code{FALSE} (default), weights are based on absolute positions. For
example, the weights of \eqn{x_0} and \eqn{x_2} used in calculating the
final weighted average of (\eqn{x_0}, null, \eqn{x_2}) are
\eqn{(1-\alpha)^2} and \eqn{1} if \code{adjust = TRUE}, and \eqn{(1-\alpha)^2}
and \eqn{\alpha} if \code{adjust = FALSE}.
\item when \code{TRUE}, weights are based on relative positions. For example, the
weights of \eqn{x_0} and \eqn{x_2} used in calculating the final weighted
average of (\eqn{x_0}, null, \eqn{x_2}) are \eqn{1-\alpha} and \eqn{1} if
\code{adjust = TRUE}, and \eqn{1-\alpha} and \eqn{\alpha} if \code{adjust = FALSE}.
}}
}
\value{
A polars \link{expression}
}
\description{
Compute exponentially-weighted moving variance
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(pl$col("a")$ewm_var(com = 1, ignore_nulls = FALSE))
}
