% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__fill_null}
\alias{expr__fill_null}
\title{Fill floating point null value with a fill value}
\usage{
expr__fill_null(value, strategy = NULL, limit = NULL)
}
\arguments{
\item{value}{Value used to fill null values.}

\item{strategy}{Strategy used to fill null values. Must be one of
\code{"forward"}, \code{"backward"}, \code{"min"}, \code{"max"}, \code{"mean"}, \code{"zero"}, \code{"one"}.}

\item{limit}{Number of consecutive null values to fill when using the
\code{"forward"} or \code{"backward"} strategy.}
}
\value{
A polars \link{expression}
}
\description{
Fill floating point null value with a fill value
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 2, NaN))
df$with_columns(
  filled_null = pl$col("a")$fill_null(strategy = "zero")
)
df$with_columns(
  filled_null = pl$col("a")$fill_null(99)
)
df$with_columns(
  filled_null = pl$col("a")$fill_null(strategy = "forward")
)
df$with_columns(
  filled_null = pl$col("a")$fill_null(pl$col("a")$median())
)
}
