% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__floor_div}
\alias{expr__floor_div}
\title{Floor divide using two expressions}
\usage{
expr__floor_div(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
\link[=expr__class]{Expr}
}
\description{
Method equivalent of floor division operator \code{expr \%/\% other}.
}
\examples{
df <- pl$DataFrame(x = 1:5)

df$with_columns(
  `x/2` = pl$col("x")$div(2),
  `x\%/\%2` = pl$col("x")$floor_div(2)
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
\item \code{\link[=expr__truediv]{<Expr>$div()}}
\item \code{\link[=expr__mod]{<Expr>$mod()}}
}
}
