% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__gt}
\alias{expr__gt}
\title{Check greater or equal inequality}
\usage{
expr__gt(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=expr__class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(2) >= 2
pl$lit(2) >= pl$lit(2)
pl$lit(2)$gt_eq(pl$lit(2))
}
