% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_finite}
\alias{expr__is_finite}
\title{Check if elements are finite}
\usage{
expr__is_finite()
}
\value{
A polars \link{expression}
}
\description{
Returns a boolean Series indicating which values are finite.
}
\examples{
pl$DataFrame(alice = c(0, NaN, NA, Inf, -Inf))$
  with_columns(finite = pl$col("alice")$is_finite())
}
