% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_infinite}
\alias{expr__is_infinite}
\title{Check if elements are infinite}
\usage{
expr__is_infinite()
}
\value{
A polars \link{expression}
}
\description{
Returns a boolean Series indicating which values are infinite.
}
\examples{
pl$DataFrame(alice = c(0, NaN, NA, Inf, -Inf))$
  with_columns(infinite = pl$col("alice")$is_infinite())
}
