% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_nan}
\alias{expr__is_nan}
\title{Check if elements are NaN}
\usage{
expr__is_nan()
}
\value{
A polars \link{expression}
}
\description{
Returns a boolean Series indicating which values are NaN.
}
\examples{
pl$DataFrame(alice = c(0, NaN, NA, Inf, -Inf))$
  with_columns(nan = pl$col("alice")$is_nan())
}
