% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_not_nan}
\alias{expr__is_not_nan}
\title{Check if elements are not NaN}
\usage{
expr__is_not_nan()
}
\value{
A polars \link{expression}
}
\description{
Returns a boolean Series indicating which values are not NaN. Syntactic sugar
for \verb{$is_nan()$not()}.
}
\examples{
pl$DataFrame(alice = c(0, NaN, NA, Inf, -Inf))$
  with_columns(not_nan = pl$col("alice")$is_not_nan())
}
