% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_not_null}
\alias{expr__is_not_null}
\title{Check if elements are NULL}
\usage{
expr__is_not_null()
}
\value{
A polars \link{expression}
}
\description{
Returns a boolean Series indicating which values are null.
}
\examples{
pl$DataFrame(x = c(1, NA, 3))$select(pl$col("x")$is_null())
}
