% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__lt}
\alias{expr__lt}
\title{Check strictly lower inequality}
\usage{
expr__lt(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=expr__class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(5) < 10
pl$lit(5) < pl$lit(10)
pl$lit(5)$lt(pl$lit(10))
}
