% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__lt_eq}
\alias{expr__lt_eq}
\title{Check lower or equal inequality}
\usage{
expr__lt_eq(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(2) <= 2
pl$lit(2) <= pl$lit(2)
pl$lit(2)$lt_eq(pl$lit(2))
}
