% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__max}
\alias{expr__max}
\title{Get maximum value}
\usage{
expr__max()
}
\value{
A polars \link{expression}
}
\description{
Get maximum value
}
\examples{
pl$DataFrame(x = c(1, NA, 3))$
  with_columns(max = pl$col("x")$max())
}
