% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__min}
\alias{expr__min}
\title{Get minimum value}
\usage{
expr__min()
}
\value{
A polars \link{expression}
}
\description{
Get minimum value
}
\examples{
pl$DataFrame(x = c(1, NA, 3))$
  with_columns(min = pl$col("x")$min())
}
