% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__neq_missing}
\alias{expr__neq_missing}
\title{Check inequality without \code{null} propagation}
\usage{
expr__neq_missing(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
df <- pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  neq = pl$col("x")$neq("y"),
  neq_missing = pl$col("x")$neq_missing("y")
)
}
\seealso{
\link{expr__neq}
}
