% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__not}
\alias{expr__not}
\title{Negate a boolean expression}
\usage{
expr__not()
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of negation operator \code{!expr}.
}
\examples{
# two syntaxes same result
pl$lit(TRUE)$not()
!pl$lit(TRUE)
}
