% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__or}
\alias{expr__or}
\title{Apply logical OR on two expressions}
\usage{
expr__or(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=expr__class]{Expr}
}
\description{
Combine two boolean expressions with OR.
}
\examples{
pl$lit(TRUE) | FALSE
pl$lit(TRUE)$or(pl$lit(TRUE))
}
