% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__reverse}
\alias{expr__reverse}
\title{Reverse a variable}
\usage{
expr__reverse()
}
\value{
A polars \link{expression}
}
\description{
Reverse a variable
}
\examples{
pl$DataFrame(a = 1:5)$select(pl$col("a")$reverse())
}
