% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__skew}
\alias{expr__skew}
\title{Compute the skewness}
\usage{
expr__skew(..., bias = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}
}
\value{
A polars \link{expression}
}
\description{
For normally distributed data, the skewness should be about zero. For
unimodal continuous distributions, a skewness value greater than zero means
that there is more weight in the right tail of the distribution.
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 3, 2, 1))
df$select(pl$col("x")$skew())
}
