% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sort}
\alias{expr__sort}
\title{Sort an Expr}
\usage{
expr__sort(..., descending = FALSE, nulls_last = FALSE)
}
\value{
A polars \link{expression}
}
\description{
Sort this column. If used in a groupby context, the groups are sorted.
}
\examples{
pl$DataFrame(a = c(6, 1, 0, NA, Inf, NaN))$
  with_columns(sorted = pl$col("a")$sort())
}
