% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_any}
\alias{expr_arr_any}
\title{Evaluate whether any boolean values in an array are true}
\usage{
expr_arr_any()
}
\value{
A polars \link{expression}
}
\description{
Evaluate whether any boolean values in an array are true
}
\examples{
df <- pl$DataFrame(
  values = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), c(NA, NA)),
  schema = list(values = pl$Array(pl$Boolean, 2))
)
df$with_columns(any = pl$col("values")$arr$any())
}
