% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_arg_min}
\alias{expr_arr_arg_min}
\title{Get the index of the minimal value in an array}
\usage{
expr_arr_arg_min()
}
\value{
A polars \link{expression}
}
\description{
Get the index of the minimal value in an array
}
\examples{
df <- pl$DataFrame(
  values = list(1:2, 2:1),
  schema = list(values = pl$Array(pl$Int32, 2))
)
df$with_columns(
  arg_min = pl$col("values")$arr$arg_min()
)
}
