% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_max}
\alias{expr_arr_max}
\title{Find the maximum value in an array}
\usage{
expr_arr_max()
}
\value{
A polars \link{expression}
}
\description{
Find the maximum value in an array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA, NA))
)$cast(pl$Array(pl$Float64, 2))
df$with_columns(max = pl$col("values")$arr$max())
}
