% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_min}
\alias{expr_arr_min}
\title{Find the minimum value in an array}
\usage{
expr_arr_min()
}
\value{
A polars \link{expression}
}
\description{
Find the minimum value in an array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA_real_, NA_real_)),
  schema = list(values = pl$Array(pl$Float64, 2))
)
df$with_columns(min = pl$col("values")$arr$min())
}
