% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_shift}
\alias{expr_arr_shift}
\title{Shift array values by \code{n} indices}
\usage{
expr_arr_shift(n = 1)
}
\value{
A polars \link{expression}
}
\description{
Shift array values by \code{n} indices
}
\examples{
df <- pl$DataFrame(
  values = list(1:3, c(2L, NA_integer_, 5L)),
  idx = 1:2,
$cast(pl$Array(pl$Int32, 3))
df$with_columns(
  shift_by_expr = pl$col("values")$arr$shift(pl$col("idx")),
  shift_by_lit = pl$col("values")$arr$shift(2)
)
}
