% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_std}
\alias{expr_arr_std}
\title{Find the standard deviation in an array}
\usage{
expr_arr_std(ddof = 1)
}
\value{
A polars \link{expression}
}
\description{
Find the standard deviation in an array
}
\examples{
df <- pl$DataFrame(
  values = list(c(2, 1, 4), c(8.4, 3.2, 1)),
  schema = list(values = pl$Array(pl$Float64, 3))
)
df$with_columns(std = pl$col("values")$arr$std())
}
