% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_to_struct}
\alias{expr_arr_to_struct}
\title{Convert array to struct}
\usage{
expr_arr_to_struct(fields = NULL)
}
\value{
A polars \link{expression}
}
\description{
Convert array to struct
}
\examples{
df <- pl$DataFrame(
  values = list(1:3, c(2L, NA_integer_, 5L)),
  schema = list(values = pl$Array(pl$Int32, 3))
)
df$with_columns(
  struct = pl$col("values")$arr$to_struct()
)

# pass a custom function that will name all fields by adding a prefix
df2 <- df$with_columns(
  pl$col("values")$arr$to_struct(
    fields = \(idx) paste0("col_", idx)
  )
)
df2

df2$unnest()
}
