% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_unique}
\alias{expr_arr_unique}
\title{Get unique values in an array}
\usage{
expr_arr_unique(maintain_order = FALSE)
}
\value{
A polars \link{expression}
}
\description{
Get unique values in an array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 1, 2), c(4, 4, 4), c(NA_real_, 6, 7)),
  schema = list(values = pl$Array(pl$Float64, 3))
)
df$with_columns(unique = pl$col("values")$arr$unique())
}
