% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_var}
\alias{expr_arr_var}
\title{Find the variance in an array}
\usage{
expr_arr_var(ddof = 1)
}
\value{
A polars \link{expression}
}
\description{
Find the variance in an array
}
\examples{
df <- pl$DataFrame(
  values = list(c(2, 1, 4), c(8.4, 3.2, 1)),
)$cast(pl$Array(pl$Float64, 3))
df$with_columns(var = pl$col("values")$arr$var())
}
