% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_cast_time_unit}
\alias{expr_dt_cast_time_unit}
\title{Change time unit}
\usage{
expr_dt_cast_time_unit(time_unit = c("ns", "us", "ms"))
}
\arguments{
\item{time_unit}{Time unit, one of 'ns', 'us', or 'ms'.}
}
\value{
A polars \link{expression}
}
\description{
Cast the underlying data to another time unit. This may lose precision.
The corresponding global timepoint will stay unchanged +/- precision.
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s"
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$dt$cast_time_unit()$alias("cast_time_unit_ns"),
  pl$col("date")$dt$cast_time_unit(time_unit = "ms")$alias("cast_time_unit_ms")
)
}
