% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_millisecond}
\alias{expr_dt_millisecond}
\title{Extract milliseconds from underlying Datetime representation}
\usage{
expr_dt_millisecond()
}
\value{
\link[=Expr_class]{Expr} of data type Int32
}
\description{
Applies to Datetime columns.
}
\examples{
df <- pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)

df$with_columns(
  millisecond = pl$col("datetime")$dt$millisecond()
)
}
