% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_minute}
\alias{expr_dt_minute}
\title{Extract minute from underlying Datetime representation}
\usage{
expr_dt_minute()
}
\value{
A polars \link{expression}
}
\description{
Returns the minute number from 0 to 59.
}
\examples{
df <- pl$DataFrame(
  date = pl$datetime_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d5s",
    time_zone = "GMT"
  )
)
df$with_columns(
  pl$col("date")$dt$minute()$alias("minute")
)
}
