% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_ordinal_day}
\alias{expr_dt_ordinal_day}
\title{Ordinal Day}
\usage{
expr_dt_ordinal_day()
}
\value{
A polars \link{expression}
}
\description{
Extract ordinal day from underlying Date representation.
Applies to Date and Datetime columns.
Returns the day of year starting from 1.
The return value ranges from 1 to 366. (The last day of year differs by years.)
}
\examples{
df <- pl$DataFrame(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d",
    time_zone = "GMT"
  )
)
df$with_columns(
  pl$col("date")$dt$ordinal_day()$alias("ordinal_day")
)
}
