% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_timestamp}
\alias{expr_dt_timestamp}
\title{Get timestamp of given Datetime}
\usage{
expr_dt_timestamp(tu = "ns")
}
\arguments{
\item{tu}{string option either 'ns', 'us', or 'ms'}
}
\value{
A polars \link{expression}
}
\description{
Return a timestamp in the given time unit.
}
\examples{
df <- pl$DataFrame(
  date = pl$datetime_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s"
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$dt$timestamp()$alias("timestamp_ns"),
  pl$col("date")$dt$timestamp(tu = "ms")$alias("timestamp_ms")
)
}
