% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_total_days}
\alias{expr_dt_total_days}
\title{Days}
\usage{
expr_dt_total_days()
}
\value{
Expr of i64
}
\description{
Extract the days from a Duration type.
}
\examples{
df <- pl$DataFrame(
  date = pl$datetime_range(
    start = as.Date("2020-3-1"),
    end = as.Date("2020-5-1"),
    interval = "1mo1s"
  )
)
df$select(
  pl$col("date"),
  diff_days = pl$col("date")$diff()$dt$total_days()
)
}
