% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_with_time_unit}
\alias{expr_dt_with_time_unit}
\title{with_time_unit}
\usage{
expr_dt_with_time_unit(time_unit = c("ns", "us", "ms"))
}
\arguments{
\item{time_unit}{Time unit, one of 'ns', 'us', or 'ms'.}
}
\value{
A polars \link{expression}
}
\description{
Set time unit of a Series of dtype Datetime or Duration.
This does not modify underlying data, and should be used to fix an incorrect time unit.
The corresponding global timepoint will change.
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s"
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$dt$with_time_unit()$alias("with_time_unit_ns"),
  pl$col("date")$dt$with_time_unit(time_unit = "ms")$alias("with_time_unit_ms")
)
}
