% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_year}
\alias{expr_dt_year}
\title{Year}
\usage{
expr_dt_year()
}
\value{
A polars \link{expression}
}
\description{
Extract year from underlying Date representation.
Applies to Date and Datetime columns.
Returns the year number in the calendar date.
}
\examples{
df <- pl$DataFrame(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d",
    time_zone = "GMT"
  )
)
df$with_columns(
  pl$col("date")$dt$year()$alias("year"),
  pl$col("date")$dt$iso_year()$alias("iso_year")
)
}
