% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_show_graph}
\alias{expr_meta_show_graph}
\title{Format the expression as a Graphviz graph}
\usage{
expr_meta_show_graph(..., show = TRUE, output_path = NULL, raw_output = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{show}{Show the figure.}

\item{output_path}{Write the figure to disk.}

\item{raw_output}{Return dot syntax. This cannot be combined with \code{show} and
/ or \code{output_path}.}
}
\value{
If \code{show = TRUE}, the same output as \code{DiagrammeR::grViz()}.
If \code{raw_output = TRUE}, a string containing the GraphViz code for the graph.
}
\description{
Note that Graphviz must be installed to render the visualization (if not
already present, you can download it here:
\verb{<https://graphviz.org/download>}_).
}
\examples{
\dontshow{if (requireNamespace("DiagrammeR", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
my_expr <- (pl$col("foo") * pl$col("bar"))$sum()$over(pl$col("ham")) / 2
my_expr$meta$show_graph()
\dontshow{\}) # examplesIf}
}
