% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_polars_dtype.R
\name{infer_polars_dtype}
\alias{infer_polars_dtype}
\alias{infer_polars_dtype.default}
\alias{infer_polars_dtype.polars_series}
\alias{infer_polars_dtype.polars_data_frame}
\alias{infer_polars_dtype.polars_lazy_frame}
\alias{infer_polars_dtype.POSIXlt}
\alias{infer_polars_dtype.NULL}
\alias{infer_polars_dtype.list}
\alias{infer_polars_dtype.AsIs}
\alias{infer_polars_dtype.data.frame}
\alias{infer_polars_dtype.vctrs_vctr}
\title{Infer polars data type corresponding to given R object}
\usage{
infer_polars_dtype(x, ...)

\method{infer_polars_dtype}{default}(x, ...)

\method{infer_polars_dtype}{polars_series}(x, ...)

\method{infer_polars_dtype}{polars_data_frame}(x, ...)

\method{infer_polars_dtype}{polars_lazy_frame}(x, ...)

\method{infer_polars_dtype}{POSIXlt}(x, ...)

\method{infer_polars_dtype}{`NULL`}(x, ...)

\method{infer_polars_dtype}{list}(x, ..., strict = FALSE)

\method{infer_polars_dtype}{AsIs}(x, ...)

\method{infer_polars_dtype}{data.frame}(x, ...)

\method{infer_polars_dtype}{vctrs_vctr}(x, ...)
}
\description{
Infer polars data type corresponding to given R object
}
