% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-ipc-functions.R
\name{pl__scan_ipc}
\alias{pl__scan_ipc}
\title{Lazily read from an Arrow IPC (Feather v2) file or multiple files via glob
patterns}
\usage{
pl__scan_ipc(
  source,
  ...,
  n_rows = NULL,
  cache = TRUE,
  rechunk = FALSE,
  row_index_name = NULL,
  row_index_offset = 0L,
  storage_options = NULL,
  memory_map = TRUE,
  retries = 2,
  file_cache_ttl = NULL,
  hive_partitioning = NULL,
  hive_schema = NULL,
  try_parse_hive_dates = TRUE,
  include_file_paths = NULL
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{memory_map}{A logical. If \code{TRUE}, try to memory map the file. This can
greatly improve performance on repeated queries as the OS may cache pages.
Only uncompressed Arrow IPC files can be memory mapped.}

\item{hive_partitioning}{Infer statistics and schema from Hive partitioned
sources and use them to prune reads. If \code{NULL} (default), it is automatically
enabled when a single directory is passed, and otherwise disabled.}

\item{hive_schema}{A list containing the column names and data types of the
columns by which the data is partitioned, e.g.
\code{list(a = pl$String, b = pl$Float32)}. If \code{NULL} (default), the schema of
the Hive partitions is inferred.}

\item{try_parse_hive_dates}{Whether to try parsing hive values as date /
datetime types.}

\item{include_file_paths}{Character value indicating the column name that
will include the path of the source file(s).}
}
\description{
This allows the query optimizer to push down predicates and projections to
the scan level, thereby potentially reducing memory overhead.
}
\details{
If \code{memory_map} is set, the bytes on disk are mapped 1:1 to memory. That
means that you cannot write to the same filename, e.g.
\code{pl$read_ipc("my_file.arrow")$write_ipc("my_file.arrow")} will fail.
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE) && arrow::arrow_with_dataset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
temp_dir <- tempfile()
# Write a hive-style partitioned arrow file dataset
arrow::write_dataset(
  mtcars,
  temp_dir,
  partitioning = c("cyl", "gear"),
  format = "arrow",
  hive_style = TRUE
)
list.files(temp_dir, recursive = TRUE)

# If the path is a folder, Polars automatically tries to detect partitions
# and includes them in the output
pl$scan_ipc(temp_dir)$collect()

# We can also impose a schema to the partition
pl$scan_ipc(temp_dir, hive_schema = list(cyl = pl$String, gear = pl$Int32))$collect()
\dontshow{\}) # examplesIf}
}
