% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{is_selector}
\alias{is_selector}
\title{Indicate whether the given object/expression is a selector}
\usage{
is_selector(x)
}
\value{
A logical value
}
\description{
Indicate whether the given object/expression is a selector
}
\examples{
is_selector(pl$col("colx"))
is_selector(!cs$integer())
}
