% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__limit}
\alias{lazyframe__limit}
\title{Get the first \code{n} rows}
\usage{
lazyframe__limit(n = 5)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Alias for \code{\link[=lazyframe__head]{<LazyFrame>$head()}}.
}
\examples{
lf <- pl$LazyFrame(a = 1:6, b = 7:12)
lf$limit()$collect()
lf$limit(2)$collect()
}
