% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-various.R
\name{parquet_statistics}
\alias{parquet_statistics}
\title{Prepare statistics for writing to Parquet file}
\usage{
parquet_statistics(
  min = TRUE,
  max = TRUE,
  distinct_count = TRUE,
  null_count = TRUE
)
}
\arguments{
\item{min}{Include stats on the minimum values in the column.}

\item{max}{Include stats on the maximum values in the column.}

\item{distinct_count}{Include stats on the number of distinct values in the
column.}

\item{null_count}{Include stats on the number of null values in the column.}
}
\description{
Prepare statistics for writing to Parquet file
}
