% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregation-horizontal.R
\name{pl__any_horizontal}
\alias{pl__any_horizontal}
\title{Apply the OR logical horizontally across columns}
\usage{
pl__any_horizontal(...)
}
\value{
A polars \link{expression}
}
\description{
Apply the OR logical horizontally across columns
}
\examples{
df <- pl$DataFrame(
  a = c(FALSE, FALSE, TRUE, TRUE, FALSE, NA),
  b = c(FALSE, TRUE, TRUE, NA, NA, NA),
  c = c("u", "v", "w", "x", "y", "z")
)

df$with_columns(
  any = pl$any_horizontal("a", "b", "c")
)
}
