% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{pl__deserialize_lf}
\alias{pl__deserialize_lf}
\title{Read a logical plan from a file to construct a LazyFrame}
\usage{
pl__deserialize_lf(source)
}
\arguments{
\item{source}{String containing the LazyFrame logical plan in JSON format.}
}
\value{
A character value
}
\description{
Read a logical plan from a file to construct a LazyFrame
}
\examples{
lf <- pl$LazyFrame(a = 1:3)$sum()
ser <- lf$serialize()
pl$deserialize_lf(ser)
}
