% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-s3-operators.R
\name{S3_arithmetic}
\alias{S3_arithmetic}
\alias{+.polars_expr}
\alias{-.polars_expr}
\alias{*.polars_expr}
\alias{/.polars_expr}
\alias{\%\%.polars_expr}
\alias{\%/\%.polars_expr}
\alias{^.polars_expr}
\alias{<.polars_expr}
\alias{>.polars_expr}
\alias{<=.polars_expr}
\alias{>=.polars_expr}
\alias{==.polars_expr}
\alias{!=.polars_expr}
\alias{!.polars_expr}
\alias{&.polars_expr}
\alias{|.polars_expr}
\title{Arithmetic operators for Polars objects}
\usage{
\method{+}{polars_expr}(e1, e2)

\method{-}{polars_expr}(e1, e2)

\method{*}{polars_expr}(e1, e2)

\method{/}{polars_expr}(e1, e2)

\method{\%\%}{polars_expr}(e1, e2)

\method{\%/\%}{polars_expr}(e1, e2)

\method{^}{polars_expr}(e1, e2)

\method{<}{polars_expr}(e1, e2)

\method{>}{polars_expr}(e1, e2)

\method{<=}{polars_expr}(e1, e2)

\method{>=}{polars_expr}(e1, e2)

\method{==}{polars_expr}(e1, e2)

\method{!=}{polars_expr}(e1, e2)

\method{!}{polars_expr}(e1)

\method{&}{polars_expr}(e1, e2)

\method{|}{polars_expr}(e1, e2)
}
\arguments{
\item{x, y}{numeric type of Polars objects or objects that can be coerced
such. Only \code{+} accepts two string inputs.}
}
\value{
A Polars object the same type as the input.
}
\description{
Arithmetic operators for Polars objects
}
\examples{
pl$lit(5) + 10
5 + pl$lit(10)
pl$lit(5) + pl$lit(10)
+pl$lit(1)

# This will not raise an error as it is not actually evaluated.
expr = pl$lit(5) + "10"
expr

# Will raise an error as it is evaluated.
tryCatch(
  pl$select(expr),
  error = function(e) e
)

# `+` accepts two string inputs
pl$select(pl$lit("a") + "b")

as_polars_series(5) + 10
+as_polars_series(5)
-as_polars_series(5)
}
\seealso{
\itemize{
\item \code{\link[=expr__add]{<Expr>$add()}}
\item \code{\link[=expr__sub]{<Expr>$sub()}}
\item \code{\link[=expr__mul]{<Expr>$mul()}}
\item \code{\link[=expr__true_div]{<Expr>$true_div()}}
\item \code{\link[=expr__pow]{<Expr>$pow()}}
\item \code{\link[=expr__mod]{<Expr>$mod()}}
\item \code{\link[=expr__floor_div]{<Expr>$floor_div()}}
\item \code{\link[=series__add]{<Series>$add()}}
\item \code{\link[=series__sub]{<Series>$sub()}}
\item \code{\link[=series__mul]{<Series>$mul()}}
\item \code{\link[=series__true_div]{<Series>$true_div()}}
\item \code{\link[=series__pow]{<Series>$pow()}}
\item \code{\link[=series__mod]{<Series>$mod()}}
\item \code{\link[=series__floor_div]{<Series>$floor_div()}}
}
}
