% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__alias}
\alias{series__alias}
\title{Rename the series}
\usage{
series__alias(name)
}
\arguments{
\item{name}{The new name.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Rename the series
}
\examples{
series <- pl$Series("a", 1:3)
series$alias("b")
}
