% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__rechunk}
\alias{series__rechunk}
\title{Create a single chunk of memory for this Series}
\usage{
series__rechunk()
}
\value{
A \link[=Series]{polars Series}
}
\description{
Create a single chunk of memory for this Series
}
\examples{
s <- pl$Series("a", c(1, 2, 3))
s$n_chunks()

s2 <- pl$Series("a", c(4, 5, 6))
s <- pl$concat(s, s2, rechunk = FALSE)
s$n_chunks()

s$rechunk()$n_chunks()
}
