% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__eq_missing}
\alias{expr__eq_missing}
\title{Check equality without \code{null} propagation}
\usage{
expr__eq_missing(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
df <- pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  eq = pl$col("x")$eq("y"),
  eq_missing = pl$col("x")$eq_missing("y")
)
}
\seealso{
\link{expr__eq}
}
