% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__lt}
\alias{expr__lt}
\title{Check strictly lower inequality}
\usage{
expr__lt(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(5) < 10
pl$lit(5) < pl$lit(10)
pl$lit(5)$lt(pl$lit(10))
}
