% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__neq}
\alias{expr__neq}
\title{Check inequality}
\usage{
expr__neq(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=expr__class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(1) != 2
pl$lit(1) != pl$lit(2)
pl$lit(1)$neq(pl$lit(2))
}
\seealso{
\link{expr__neq_missing}
}
