% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__skew}
\alias{expr__skew}
\title{Compute the skewness}
\usage{
expr__skew(..., bias = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}
}
\value{
A polars \link{expression}
}
\description{
For normally distributed data, the skewness should be about zero. For
unimodal continuous distributions, a skewness value greater than zero means
that there is more weight in the right tail of the distribution.
}
\details{
The sample skewness is computed as the Fisher-Pearson coefficient of
skewness, i.e.
$g_1=\frac{m_3}{m_2^{3/2}}$
where
$m_i=\frac{1}{N}\sum_{n=1}^N(x\link{n}-\bar{x})^i$
is the biased sample $i\texttt{th}$ central moment, and $\bar{x}$ is the
sample mean. If \code{bias} is \code{FALSE}, the calculations are corrected for bias
and the value computed is the adjusted Fisher-Pearson standardized moment
coefficient, i.e.
$G_1 = \frac{k_3}{k_2^{3/2}} = \frac{\sqrt{N(N-1)}}{N-2}\frac{m_3}{m_2^{3/2}}$
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 3, 2, 1))
df$select(pl$col("x")$skew())
}
