% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_with_time_unit}
\alias{expr_dt_with_time_unit}
\title{Set time unit of a Series of dtype Datetime or Duration}
\usage{
expr_dt_with_time_unit(time_unit = c("ns", "us", "ms"))
}
\arguments{
\item{time_unit}{Time unit, one of 'ns', 'us', or 'ms'.}
}
\value{
A polars \link{expression}
}
\description{
This is deprecated. Cast to Int64 and then to Datetime instead.
}
